dress_modifier/requirements.txt
diffusers==0.29.2
transformers==4.41.2
torch==2.3.1
torchvision==0.18.1
gradio==4.38.1
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==1.26.4
accelerate==0.32.1